+++ packages/apollo-angular-link-http-common/ng-package.js
@@ -1,0 +1,10 @@
+const {umdModuleIds} = require('../../ng-package');
+
+module.exports = {
+  dest: 'build',
+  lib: {
+    entryFile: 'src/index.ts',
+    flatModuleFile: 'ng.apolloLink.httpCommon',
+    umdModuleIds,
+  },
+};