+++ packages/apollo-angular/ng-package.js
@@ -1,0 +1,13 @@
+const {umdModuleIds} = require('../../ng-package');
+
+module.exports = {
+  dest: 'build',
+  lib: {
+    entryFile: 'src/index.ts',
+    flatModuleFile: 'ng.apollo',
+    umdModuleIds,
+  },
+  whitelistedNonPeerDependencies: [
+    "semver"
+  ]
+};