+++ packages/apollo-angular/rollup.config.js
@@ -1,3 +1,0 @@
-import build from '../../rollup.config';
-
-export default build('core');