+++ packages/apollo-angular/src/index.ts
@@ -1,5 +1,9 @@
 export {Apollo, ApolloBase} from './Apollo';
 export {QueryRef} from './QueryRef';
+export {Query} from './Query';
+export {Mutation} from './Mutation';
+export {Subscription} from './Subscription';
 export {SelectPipe} from './SelectPipe';
 export {ApolloModule} from './ApolloModule';
 export {APOLLO_OPTIONS} from './tokens';
+export {SubscriptionResult} from './types';