+++ packages/apollo-angular/testing/tests/integration.spec.ts
@@ -1,10 +1,17 @@
 import {setupAngular} from './_setup';
 import {Apollo} from 'apollo-angular';
 import {TestBed, inject} from '@angular/core/testing';
+import {addTypenameToDocument} from 'apollo-utilities';
+import {print} from 'graphql';
 
 import gql from 'graphql-tag';
 
-import {ApolloTestingModule, ApolloTestingController} from '../src';
+import {
+  ApolloTestingModule,
+  ApolloTestingController,
+  APOLLO_TESTING_CACHE,
+} from '../src';
+import {InMemoryCache} from 'apollo-cache-inmemory';
 
 describe('Integration', () => {
   let apollo: Apollo;
@@ -18,15 +25,13 @@ describe('Integration', () => {
     });
   });
 
-  beforeEach(
-    inject(
-      [Apollo, ApolloTestingController],
-      (_apollo: Apollo, _backend: ApolloTestingController) => {
-        apollo = _apollo;
-        backend = _backend;
-      },
-    ),
-  );
+  beforeEach(inject(
+    [Apollo, ApolloTestingController],
+    (_apollo: Apollo, _backend: ApolloTestingController) => {
+      apollo = _apollo;
+      backend = _backend;
+    },
+  ));
 
   afterEach(() => {
     backend.verify();
@@ -166,12 +171,63 @@ describe('Integration', () => {
 
     backend
       .expectOne(operation => {
-        expect(operation.operationName).toEqual(op.operationName);
+        expect(operation.operationName).toBe(op.operationName);
         expect(operation.variables).toEqual(op.variables);
-        expect(operation.query).toEqual(op.query);
+        expect(print(operation.query)).toBe(print(op.query));
 
         return true;
       })
       .flush({data});
+  });
+
+  test('it should be able to test with fragments', done => {
+    TestBed.resetTestingModule();
+    TestBed.configureTestingModule({
+      imports: [ApolloTestingModule],
+      providers: [
+        {
+          provide: APOLLO_TESTING_CACHE,
+          useValue: new InMemoryCache({addTypename: true}),
+        },
+      ],
+    });
+
+    const apollo: Apollo = TestBed.get(Apollo);
+    const backend: ApolloTestingController = TestBed.get(
+      ApolloTestingController,
+    );
+
+    const query = gql`
+      {
+        heroes {
+          ...fields
+        }
+      }
+      fragment fields on Character {
+        name
+      }
+    `;
+
+    const op = {
+      query,
+    };
+
+    const data = {
+      heroes: [
+        {
+          name: 'Superman',
+          __typename: 'Character',
+        },
+      ],
+    };
+
+    apollo.query<any>(op).subscribe({
+      next: result => {
+        expect(result.data).toMatchObject(data);
+        done();
+      },
+    });
+
+    backend.expectOne(addTypenameToDocument(query)).flush({data});
   });
 });