+++ packages/apollo-angular-boost/CHANGELOG.md
@@ -1,6 +1,22 @@
 # Change log
 
-### vNEXT
+### vNEXR
+
+### v1.1.0
+
+- Bump `apollo-angular` to v1.3.0
+
+### v1.0.0
+
+- Uses `ng-packagr`
+
+### v1.0.0-beta.4
+
+- Exports all from `apollo-angular`
+
+### v1.0.0-beta.2, ### v1.0.0-beta.3
+
+- Fixes an issue with initialization via `APOLLO_BOOST_CONFIG`
 
 ### v1.0.0-beta.1